Changing some fix for better results
What was previously happening was the target was changed to
target + "/" + idir
where idir was
/wp-content/plugins/
so ex if I was testing on : https://hostelgate.lt/ ----> python3 nettacker.py -i https://hostelgate.lt/ -m wp_plugin_scan
there is '/' at the end of the target and there is already '/' in the idir variable and an extra '/' is added in the target variable when calling check() function. So the final url was
https://hostelgate.lt///wp-content/plugins/{}
So I removed them to improve the results. Now it gives:
https://hostelgate.lt/wp-content/plugins/{}
This I found when testing on a website and the tool was giving some false positives.